move loading into its own directory
