+++ src/components/index.js
@@ -7,5 +7,7 @@ export {default as EmptyEvents} from './EmptyEvents/EmptyEvents';
 export {default as EventDetails} from './EventDetails/EventDetails';
 export {default as EventsList} from './EventsList/EventsList';
 export {default as EventsListItem} from './EventsListItem/EventsListItem';
+
+export {default as Loading} from './Loading/Loading';
 export {default as SignIn} from './SignIn/SignIn';
 export {default as SignUp} from './SignUp/SignUp';
